Improve black startup to format a single file. Fixes #2564
