+++ .gitignore
@@ -24,3 +24,7 @@ src/_black_version.py
 .hypothesis/
 venv/
 .ipynb_checkpoints/
+/.mypy_cache/
+.project
+.pydevproject
+/.pytest_cache/
